refactor(studio): extract useAskAgentModal from useDomEditSession
